move section items out of main so we can compile unit tests
