+++ CMakeLists.txt
@@ -21,9 +21,16 @@ include(FetchContent)
 FetchContent_Declare(json URL https://github.com/nlohmann/json/releases/download/v3.11.3/json.tar.xz)
 FetchContent_MakeAvailable(json)
 
-add_library(rainbow_grades output.cpp table.cpp student.cpp zone.cpp benchmark.cpp submini_polls.cpp
-        grade.cpp
-        gradeable.cpp)
+add_library(rainbow_grades
+            output.cpp
+            table.cpp
+            student.cpp
+            zone.cpp
+            benchmark.cpp
+            submini_polls.cpp
+            grade.cpp
+            gradeable.cpp
+            section.cpp)
 target_link_libraries(rainbow_grades PRIVATE nlohmann_json::nlohmann_json)
 target_include_directories(rainbow_grades PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
